+++ app/src/main/java/ru/flawden/baskovmusic/ui/AppState.kt
@@ -4,6 +4,7 @@ import ru.flawden.baskovmusic.model.AccountInfo
 import ru.flawden.baskovmusic.model.GuildSummary
 import ru.flawden.baskovmusic.model.HomeSnapshot
 import ru.flawden.baskovmusic.model.LibrarySnapshot
+import ru.flawden.baskovmusic.model.LocalTrack
 import ru.flawden.baskovmusic.model.MixDetail
 import ru.flawden.baskovmusic.model.MixesSnapshot
 import ru.flawden.baskovmusic.model.TrackPreview
@@ -14,6 +15,12 @@ sealed interface AppScreen {
     data class GuildPicker(val account: AccountInfo, val guilds: List<GuildSummary>) : AppScreen
     data class Home(val account: AccountInfo, val guild: GuildSummary, val snapshot: HomeSnapshot) : AppScreen
     data class Library(val account: AccountInfo, val guild: GuildSummary, val snapshot: LibrarySnapshot) : AppScreen
+    data class LocalMusic(
+        val account: AccountInfo,
+        val guild: GuildSummary,
+        val permissionGranted: Boolean,
+        val tracks: List<LocalTrack>,
+    ) : AppScreen
     data class Mixes(val account: AccountInfo, val guild: GuildSummary, val snapshot: MixesSnapshot) : AppScreen
     data class Mix(val account: AccountInfo, val guild: GuildSummary, val detail: MixDetail) : AppScreen
     data class Track(val account: AccountInfo, val guild: GuildSummary, val track: TrackPreview) : AppScreen